feature: 인수 테스트를 충족하는 기능 구현
* Line 관련 비지니스 로직 작성
* LineAcceptanceTest에서 @Sql 사용하는 방식으로 변경
* line truncate sql파일 추가
* README 변경

+++ src/main/java/subway/line/LineRequest.java
@@ -5,7 +5,7 @@ public class LineRequest {
     private String color;
     private Long upStationId;
     private Long downStationId;
-    private int distance;
+    private Long distance;
 
     public LineRequest() {
     }
@@ -26,7 +26,7 @@ public class LineRequest {
         return downStationId;
     }
 
-    public int getDistance() {
+    public Long getDistance() {
         return distance;
     }
 }
